requirement.txt
scikit-learn
pandas
numpy
matplotlib.pyplot
sklearn.cluster
sklearn.neighbors
PIL